+++ pkg/coreapi/graph/resolvers/function_run.resolver.go
@@ -3,14 +3,49 @@ package resolvers
 import (
 	"context"
 	"encoding/json"
+	"fmt"
 	"time"
 
+	"github.com/google/uuid"
 	"github.com/inngest/inngest/pkg/coreapi/graph/models"
 	"github.com/inngest/inngest/pkg/enums"
 	"github.com/inngest/inngest/pkg/event"
 	"github.com/inngest/inngest/pkg/execution/state"
 	"github.com/oklog/ulid/v2"
 )
+
+func (r *functionRunResolver) Status(ctx context.Context, obj *models.FunctionRun) (*models.FunctionRunStatus, error) {
+	md, err := r.Runner.StateManager().Metadata(ctx, ulid.MustParse(obj.ID))
+	if err != nil {
+		return nil, fmt.Errorf("Run ID not found: %w", err)
+	}
+	status := models.FunctionRunStatusRunning
+	switch md.Status {
+	case enums.RunStatusCompleted:
+		status = models.FunctionRunStatusCompleted
+	case enums.RunStatusFailed:
+		status = models.FunctionRunStatusFailed
+	case enums.RunStatusCancelled:
+		status = models.FunctionRunStatusCancelled
+	}
+	return &status, nil
+}
+
+func (r *functionRunResolver) PendingSteps(ctx context.Context, obj *models.FunctionRun) (*int, error) {
+	md, err := r.Runner.StateManager().Metadata(ctx, ulid.MustParse(obj.ID))
+	if err != nil {
+		return nil, fmt.Errorf("Run ID not found: %w", err)
+	}
+	return &md.Pending, nil
+}
+
+func (r *functionRunResolver) Function(ctx context.Context, obj *models.FunctionRun) (*models.Function, error) {
+	fn, err := r.Data.GetFunctionByID(ctx, uuid.MustParse(obj.FunctionID))
+	if err != nil {
+		return nil, err
+	}
+	return models.MakeFunction(fn)
+}
 
 // TODO Duplicate code. Move to field-level resolvers and add dataloaders.
 func (r *functionRunResolver) Timeline(ctx context.Context, obj *models.FunctionRun) ([]models.FunctionRunEvent, error) {